apps/knowledge_base/knowledge_document_mgmt/serializers.py
from rest_framework import serializers

from apps.knowledge_base.models import KnowledgeDocument


class KnowledgeDocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = KnowledgeDocument
        fields = "__all__"
