requirements/requirements.txt
Django==4.2.7
django-split-settings==1.2.0
python-dotenv==1.0.0
daphne==4.0.0
better_exceptions==0.3.3

django-comment-migrate==0.1.7
joblib==1.3.2
pandas==2.1.4

pydantic==2.5.3
singleton-decorator==1.0.0

djangorestframework==3.14.0
django-filter==23.2
djangorestframework_simplejwt==5.3.0

django-cors-headers==4.2.0

django-celery-beat==2.5.0
django-celery-results==2.5.0
sqlalchemy==2.0.20

requests==2.31.0
python-keycloak==3.7.0

pycryptodome==3.20.0

redis==5.0.1

whitenoise==6.5.0
wrapt==1.16.0
django-auditlog==2.3.0

psycopg2-binary==2.9.9

django-import-export==3.3.6
antlr4_python3_runtime==4.11.1

#apache-age-python==0.0.7

# minio相关
filetype==1.0.13
minio==7.1.13
django-minio-backend==3.6.0